Add pagination features DWRE-10904 (#1)
Adding a new iterator that can be instantiated with a given set of entries. This is necessary for pagination since we fetch everything beforehand but do not want to change the interface of the iterator in order to keep the code changes as local as possible. Also adding params for page size and pagination.